支払方法選択&作成/購入フェーズお客様情報入力画面に組み込み完了・機能実装

+++ database/migrations/2026_06_23_082508_create_payment_methods_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('card_holder')->nullable();
             $table->string('last4', 4)->nullable();
             $table->unsignedTinyInteger('exp_month')->nullable();
-            $table->unsignedTinyInteger('exp_year')->nullable();
+            $table->unsignedSmallInteger('exp_year')->nullable();
 
             // PayPay用
             $table->string('paypay_phone')->nullable();
